Some scripts I have written over the years
fits_examine - Takes the input of a fits image, pulls out the number of pixels in x and y, and creates a grid space to determine which pixels are masked out. Outputs a file with the size of the CCD and another file with the mask positions.

GCS_likelihood - A Maximum-Likelihood Estimation script. Inputs the spatial coordinates of some object (we used the RGB stellar population of 3 dwarf galaxies and the GCS system of M86) and determines the best fitting parameters for a Sersic function or exponential function.

KNNmodule_v2 - A for-fun project to see if I could recreate one of the packages I was using at the time. Takes the spatial coordinates of some data and estimates the surface density at each data point or in a grid across the field using the Kth Nearest Neighbor surface density estimation technique.

KDEmodule - A script that estimates the surface density across the field by using Kernel Density Estimation. Takes spatial coordinates (in pixels or RADEC) and outputs a plot of the background subtracted surface density

e_r_bins - A kludgy mess that outputs the elliptical radial profile of a data set. Used to test our MLE script and determine how accurate the parameters returned were.